offline client manegment

+++ app/lib/Blocs/clients/bloc/clients_bloc.dart
@@ -1,4 +1,6 @@
 import 'dart:async';
+import 'package:app/db/db.dart';
+import 'package:app/utils/connection_checker.dart';
 import 'package:app/validation/validator.dart';
 import 'package:meta/meta.dart';
 import 'package:app/models/client.dart';
@@ -14,6 +16,8 @@ class ClientsBloc extends Bloc<ClientsEvent, ClientsState> {
   List<Client> clients = [];
   int page = 1;
   bool endOfPage = false;
+  bool syncing = false;
+  bool isFirstFetch = false;
   ClientsBloc({required this.clientsRepository}) : super(ClientsInitial());
 
   @override
@@ -42,6 +46,8 @@ class ClientsBloc extends Bloc<ClientsEvent, ClientsState> {
       yield* _mapSearchClientsEventToState(
         event.key,
       );
+    } else if (event is SyncClientEvent) {
+      syncClients();
     }
   }
 
@@ -50,6 +56,11 @@ class ClientsBloc extends Bloc<ClientsEvent, ClientsState> {
     bool loadMore,
   ) async* {
     yield ClientFetchingState();
+    if (isFirstFetch) {
+      if (!syncing) {
+        await syncClients();
+      }
+    }
     int prevPage = page;
     try {
       if (page <= 0) {
@@ -57,9 +68,11 @@ class ClientsBloc extends Bloc<ClientsEvent, ClientsState> {
       }
       if (loadMore) {
         if (endOfPage) {
+          isFirstFetch = false;
           yield ClientFetchingSuccessState(
             clients: this.clients,
           );
+
           return;
         }
         final reqData = await clientsRepository.getClients(dataX, page);
@@ -74,11 +87,13 @@ class ClientsBloc extends Bloc<ClientsEvent, ClientsState> {
           page++;
         }
       }
+      isFirstFetch = false;
       yield ClientFetchingSuccessState(
         clients: this.clients,
       );
       return;
     } catch (e) {
+      isFirstFetch = false;
       if (page > prevPage) {
         page--;
       }
@@ -111,12 +126,36 @@ class ClientsBloc extends Bloc<ClientsEvent, ClientsState> {
   Stream<ClientsState> _mapCreateClientToState(CreateEditData data) async* {
     yield ClientCreatingState();
     try {
-      Client? clientX = await clientsRepository.createClient(data);
-
       int prevPageCount = this.clients.length ~/ 5;
-      if (clientX != null) {
-        this.clients.add(clientX);
+      bool connected = await ConnectionChecker.CheckInternetConnection();
+      print("-s--connected--${connected}");
+      if (!connected) {
+        print("create---not-connected-to-internet");
+        data.type = 'CREATE';
+        CreateEditData? clientCreated = await CsvDatabse.instance.create(data);
+        if (clientCreated != null) {
+          Client clientX = Client(
+              isLocal: true,
+              addresses: clientCreated.addresses,
+              orders: [],
+              firstName: clientCreated.firstName,
+              lastName: clientCreated.lastName,
+              mobile: clientCreated.mobile,
+              email: clientCreated.email,
+              id: int.parse(
+                clientCreated.id.toString(),
+              ));
+          this.clients.add(clientX);
+        }
+      } else {
+        print("create---connected-to-internet");
+        Client? clientX = await clientsRepository.createClient(data);
+
+        if (clientX != null) {
+          this.clients.add(clientX);
+        }
       }
+
       int currentPageCount = this.clients.length ~/ 5;
       if (this.clients.length == 0) {
         page = 1;
@@ -128,6 +167,8 @@ class ClientsBloc extends Bloc<ClientsEvent, ClientsState> {
       yield ClientCreateSuccesstate();
       return;
     } catch (e) {
+      print("bloc --create--client");
+      print(e);
       yield ClientCreateFailedState(message: e.toString());
     }
   }
@@ -135,16 +176,58 @@ class ClientsBloc extends Bloc<ClientsEvent, ClientsState> {
   Stream<ClientsState> _mapUpdateClientToState(CreateEditData data) async* {
     yield ClientUpdatingState();
     try {
-      Client? clientX = await clientsRepository.updateClient(data);
-      if (clientX != null) {
-        this.clients[this
-            .clients
-            .indexWhere((element) => element.id == clientX.id)] = clientX;
-      }
+      bool connected = await ConnectionChecker.CheckInternetConnection();
+      if (!connected) {
+        print("update ---no internet");
+        if (data.id != null) {
+          data.type = 'UPDATE';
 
+          CreateEditData? storedClient = await CsvDatabse.instance
+              .readClient(int.parse(data.id.toString()));
+          if (storedClient != null) {
+            print("there is already stored client");
+            int? idDeleted = await CsvDatabse.instance
+                .deleteClient(int.parse(data.id.toString()));
+            print("so he is removed ");
+            if (idDeleted != null) {}
+          }
+
+          CreateEditData? clientCreated =
+              await CsvDatabse.instance.create(data);
+          if (clientCreated != null) {
+            print("we created again");
+            Client clientX = Client(
+                isLocal: true,
+                addresses: clientCreated.addresses,
+                orders: [],
+                firstName: clientCreated.firstName,
+                lastName: clientCreated.lastName,
+                mobile: clientCreated.mobile,
+                email: clientCreated.email,
+                id: int.parse(
+                  clientCreated.id.toString(),
+                ));
+            // ignore: todo
+            //TODO use update instead of
+            this.clients.removeWhere((element) => element.id.toString() == data.id);
+
+            this.clients.add(clientX);
+          }
+        }
+      } else {
+        print("update---connected--to--internet");
+        Client? clientX = await clientsRepository.updateClient(data);
+        if (clientX != null) {
+          this.clients[this
+              .clients
+              .indexWhere((element) => element.id == clientX.id)] = clientX;
+        }
+      }
       yield ClientUpdateSuccesstate();
       return;
     } catch (e) {
+      print("bloc --update--client");
+      print(e);
       yield ClientUpdateFailedState(message: e.toString());
     }
   }
@@ -152,7 +235,16 @@ class ClientsBloc extends Bloc<ClientsEvent, ClientsState> {
   Stream<ClientsState> _mapDeleteClientToState(String id) async* {
     yield ClientDeletingState();
     try {
-      await clientsRepository.deleteClient(id);
+      bool connected = await ConnectionChecker.CheckInternetConnection();
+      if (!connected) {
+        print("delete---not connected");
+        int? insertedID =
+            await CsvDatabse.instance.insertClientId(int.parse(id));
+        if (insertedID != null) {}
+      } else {
+        print("delete--- connected");
+        await clientsRepository.deleteClient(id);
+      }
       int prevPageCount = clients.length ~/ 5;
       print("prev -length--${this.clients.length}");
       this.clients.removeWhere((cl) => cl.id.toString() == id);
@@ -169,11 +261,77 @@ class ClientsBloc extends Bloc<ClientsEvent, ClientsState> {
       yield ClientFetchingSuccessState(
         clients: this.clients,
       );
-
       return;
     } catch (e) {
+      print("bloc --delete--client");
+      print(e);
       yield ClientDeleteFailedState(message: e.toString());
     }
+  }
+
+  Future<void> syncClients() async {
+    if (syncing) {
+      return;
+    }
+
+    syncing = true;
+    print("syncing started");
+    try {
+      print("0");
+      List<CreateEditData>? clients = await CsvDatabse.instance.readClients();
+      print("1");
+      if (clients != null) {
+        print("2");
+        for (var client in clients) {
+          print("3");
+
+          try {
+            Client? clientX = await clientsRepository.createClient(client);
+            print("delete fetched");
+            if (client.id != null) {
+              await CsvDatabse.instance
+                  .deleteClient(int.parse(client.id.toString()));
+              if (clientX != null) {
+                this.clients[this
+                        .clients
+                        .indexWhere((element) => element.id == client.id)] =
+                    clientX;
+              }
+            }
+          } catch (e) {
+            print("inner--bloc---error--sync-clients");
+            print(e);
+          }
+        }
+      }
+      print("fetching deleted client id");
+      List<int> deletedIds =
+          await CsvDatabse.instance.readAllDeletedClientIds();
+      print("deleted id");
+      print(deletedIds);
+      for (var id in deletedIds) {
+        await clientsRepository.deleteClient(id.toString());
+        CsvDatabse.instance.deleteClientID(id);
+        int prevPageCount = this.clients.length ~/ 5;
+        print("prev -length--${this.clients.length}");
+        this.clients.removeWhere((cl) => cl.id.toString() == id);
+        print("current -length--${this.clients.length}");
+
+        int currentPageCount = this.clients.length ~/ 5;
+        if (this.clients.length == 0) {
+          page = 1;
+        } else {
+          if (currentPageCount < prevPageCount) {
+            page--;
+          }
+        }
+      }
+    } catch (e) {
+      syncing = false;
+      print("bloc---error--sync-clients");
+      print(e);
+    }
+    syncing = false;
   }
 }
